Fix GoalCalculator: pass Int (not Long) to coerceAtLeast / BigDecimal
CI run 25360852761:
  e: GoalCalculator.kt:116:56 Argument type mismatch: actual Long, expected Int.

LocalDate.toEpochDays() returns Int in kotlinx-datetime 0.6.x, so
coerceAtLeast must take Int. Same for the BigDecimal constructor —
BigDecimal(Int) is fine; the .toLong() bounce was unnecessary.

+++ app/src/main/kotlin/com/fintrack/domain/goals/GoalCalculator.kt
@@ -113,8 +113,8 @@ object GoalCalculator {
         val totalDays = goal.targetDate.toEpochDays() - createdAtDate.toEpochDays()
         if (totalDays <= 0) return HUNDRED
         val elapsedDays = today.toEpochDays() - createdAtDate.toEpochDays()
-        val raw = BigDecimal(elapsedDays.coerceAtLeast(0L))
-            .divide(BigDecimal(totalDays.toLong()), 6, RoundingMode.HALF_UP)
+        val raw = BigDecimal(elapsedDays.coerceAtLeast(0))
+            .divide(BigDecimal(totalDays), 6, RoundingMode.HALF_UP)
             .multiply(HUNDRED)
             .setScale(2, RoundingMode.HALF_UP)
         return when {
